@@ -2,7 +2,7 @@
   "name": "@lasalefamine/eslint-config-base",
   "version": "3.1.1",
   "description": "Eslint JS sharable base config",
-  "main": "index.js",
+  "main": "index.mjs",
   "repository": {
     "type": "git",
     "url": "https://github.com/LasaleFamine/eslintc"
@@ -35,15 +35,15 @@
     "typescript": ">=5"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=20"
   },
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
     "confusing-browser-globals": "^1.0.11",
-    "object.assign": "^4.1.4",
-    "object.entries": "^1.1.7"
+    "object.assign": "^4.1.5",
+    "object.entries": "^1.1.8"
   },
   "gitHead": "cba38fe827537a2bac8222ae2ffb2cf8f765a968"
 }
