@@ -32,7 +32,7 @@
     "eslint-plugin-react-hooks": ">=4.6.0"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=20"
   },
   "publishConfig": {
     "access": "public"
@@ -40,8 +40,8 @@
   "dependencies": {
     "@lasalefamine/eslint-config-base": "file:../eslint-config-base",
     "confusing-browser-globals": "^1.0.11",
-    "object.assign": "^4.1.4",
-    "object.entries": "^1.1.7"
+    "object.assign": "^4.1.5",
+    "object.entries": "^1.1.8"
   },
   "gitHead": "cba38fe827537a2bac8222ae2ffb2cf8f765a968"
 }
