@@ -1,4 +1,4 @@
-module.exports = {
+export const style = [{
   rules: {
     // enforce line breaks after opening and before closing array brackets
     // https://eslint.org/docs/rules/array-bracket-newline
@@ -532,4 +532,4 @@ module.exports = {
     // require regex literals to be wrapped in parentheses
     'wrap-regex': 'off',
   },
-};
+}];
