@@ -1,6 +1,6 @@
 // const confusingBrowserGlobals = require('confusing-browser-globals');
 
-module.exports = {
+export const variables = [{
   rules: {
     // enforce or disallow variable initializations at definition
     'init-declarations': 'off',
@@ -41,4 +41,4 @@ module.exports = {
     // disallow use of variables before they are defined
     'no-use-before-define': ['error', { functions: true, classes: true, variables: true }],
   },
-};
+}];
