@@ -1,8 +1,5 @@
-module.exports = {
-  env: {
-    node: true,
-  },
 
+export const node = [{
   rules: {
     // enforce return after a callback
     'callback-return': 'off',
@@ -40,4 +37,4 @@ module.exports = {
     // disallow use of synchronous methods (off by default)
     'no-sync': 'off',
   },
-};
+}];
